// E-Store angular and node js , SQL/estore/src/app/home/types/cart.type.ts
import { Product } from "../Components/products/products.type";

export interface CartItem {
  product: Product;
  quantity: number;
  amount: number;
}

export interface Cart {
  products: CartItem[];
  totalAmount: number;
  totalProducts: number;
}


export interface DeliveryAddress {
  
  userName: string;
  address: string;
  city: string;
  state: string;
  pin: string;
}